Release V1
Includes basic styling (cuz i got lazy lol). The to do list has features such as: Checking individual to do list items, checking lists, creating, editing and deleting lists and list items with their info, etc.

Lists have a title, a description, and tasks, while list items contain a title, a description, a due date, a priority option for choosing how urgent it is, and a checklist made up of mini tasks that can be checked off when clicked on.